requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==24.1.0
beautifulsoup4==4.12.3
boto3==1.34.153
botocore==1.34.153
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cohere==5.6.2
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.3
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
executing==2.0.1
fastavro==1.9.5
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
llama-cloud==0.0.11
llama-index==0.10.59
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.59
llama-index-embeddings-cohere==0.1.9
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-cohere==0.2.0
llama-index-llms-openai==0.1.27
llama-index-llms-openai-like==0.1.3
llama-index-llms-openrouter==0.1.4
llama-index-multi-modal-llms-openai==0.1.8
llama-index-postprocessor-cohere-rerank==0.1.7
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.32
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.39.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
regex==2024.7.24
requests==2.32.3
s3transfer==0.10.2
safetensors==0.4.4
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
striprtf==0.0.26
sympy==1.12
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0+cpu
torchaudio==2.4.0+cpu
torchvision==0.19.0+cpu
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.4
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
